Month3/Week2/requirements.txt
# Core Data Science
numpy>=1.19.5
pandas>=1.2.0
scikit-learn>=0.24.2
scipy>=1.6.0

# Visualization
matplotlib>=3.3.4
seaborn>=0.11.1
plotly>=4.14.3

# Dimensionality Reduction
umap-learn>=0.5.1

# Jupyter
jupyter>=1.0.0
ipykernel>=5.4.0

# Additional Utilities
tqdm>=4.56.0
joblib>=1.0.1

# For image processing (if needed)
Pillow>=8.1.0
opencv-python>=4.5.1

# For interactive widgets
ipywidgets>=7.6.3
